#6 parse a batch line
