+++ src/nakadi/mod.rs
@@ -1,3 +1,4 @@
 mod model;
 mod components;
 mod stream;
+mod batchline;
